backend/requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database (SQLite only)
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite

# LLM and AI
google-generativeai>=0.3.0

# Web Requests and Scraping (Simplified)
httpx==0.26.0
beautifulsoup4==4.12.3
requests==2.31.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
python-json-logger==2.0.7
tenacity==8.2.3
cachetools==5.3.2

# Report Generation
jinja2==3.1.3
markdown==3.5.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
faker==22.6.0

# Code Quality
ruff==0.1.14
black==24.1.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
